Add new functionality of drawing on holding click and dragging
Each child node of the container now have an event listener which
toggles a boolean value each time a click is performed, so that, if the
click is being held, the cells can be coloured just by moving the
cursor. Once released the button, it stops colouring: an event listening
to mouseup toggles the boolean value to false. However, the simple click
still performs the color change.